src/main/java/guru/springframework/springrecipe/converters/to/model/NotesCommandToNotes.java
package guru.springframework.springrecipe.converters.to.model;

import guru.springframework.springrecipe.commands.NotesCommand;
import guru.springframework.springrecipe.domain.Notes;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

/**
 * @author kas
 */
@Component
public class NotesCommandToNotes implements Converter<NotesCommand, Notes> {


    // see: https://projectlombok.org/features/Synchronized
    @Synchronized // locks on generated field $lock and on static field $LOCK if method is static
    @Nullable
    @Override
    public Notes convert(NotesCommand notesCommand) {
        if (notesCommand == null) {
            return null;
        }
        final Notes notes = new Notes();
        notes.setId(notesCommand.getId());
        notes.setRecipeNotes(notesCommand.getRecipeNotes());
        return notes;
    }
}
